Documented tests and used ulp where possible

+++ include/nanobind_layers.hpp
@@ -12,11 +12,11 @@ namespace nb = nanobind;
 // Eigen::RowMajor>> &weight doesn't change the order of the data as a
 // referenced is passed to the data And it's quicker to transpose the data here
 // than in cpp But i'll come nack to this at the end
-Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> lin(
-    const nb::DRef<Eigen::MatrixXf> &input,
-    const nb::DRef<Eigen::MatrixXf> &weight,
-    const nb::DRef<Eigen::MatrixXf> &bias) {
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> res =
+Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> lin(
+    const nb::DRef<Eigen::MatrixXd> &input,
+    const nb::DRef<Eigen::MatrixXd> &weight,
+    const nb::DRef<Eigen::MatrixXd> &bias) {
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> res =
         (input * weight.transpose());
     for (size_t i = 0; i < res.rows(); ++i) { res.row(i) += bias.transpose(); }
     return res;
@@ -26,12 +26,12 @@ Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> lin(
 // but again I'll come back to this
 // Also note the reference to the data is passed and the data is changed in
 // place so nothing returned
-void ReLU(nb::DRef<Eigen::MatrixXf> input) {
+void ReLU(nb::DRef<Eigen::MatrixXd> input) {
     int N = input.rows();
     int M = input.cols();
     for (size_t i = 0; i < N; ++i) {
         for (size_t j = 0; j < M; ++j) {
-            input(i, j) = std::max(input(i, j), static_cast<float>(0.0));
+            input(i, j) = std::max(input(i, j), 0.0);
         }
     }
     return;
@@ -41,13 +41,12 @@ void ReLU(nb::DRef<Eigen::MatrixXf> input) {
 // but again I'll come back to this
 // Also note the reference to the data is passed and the data is changed in
 // place so nothing returned
-void sigmoid(nb::DRef<Eigen::MatrixXf> input) {
+void sigmoid(nb::DRef<Eigen::MatrixXd> input) {
     int N = input.rows();
     int M = input.cols();
     for (size_t i = 0; i < N; ++i) {
         for (size_t j = 0; j < M; ++j) {
-            input(i, j) = static_cast<float>(1.0) /
-                          (static_cast<float>(1.0) + std::exp(-input(i, j)));
+            input(i, j) = 1.0 / (1.0 + std::exp(-input(i, j)));
         }
     }
     return;
